feat: Obter funcionários do banco de dados e injeções de dependencias

+++ Distribuicao.Lucros.Domain/Entities/Funcionario.cs
@@ -8,7 +8,7 @@ namespace Distribuicao.Lucros.Domain.Entities
         public string nome { get; set; }
         public string area { get; set; }
         public string cargo { get; set; }
-        public decimal salario_bruto { get; set; }
+        public double salario_bruto { get; set; }
         public DateTime data_de_admissao { get; set; }
 
     }
